Small changes
Bumped the IDF version upon which this app is
built to the current snapshot of the release/v4.2
branch simply because it's the one I'm using
right now for another project and I want to save
disk space.

Removed the redundant pem certificate from the
mqtt component, as it's the same used on the
esp_http_client example. Now both share the same
certificate.