Buch of small changes
Disabled MQTT**S** due to the mqtt example
certificate expiration. Will re-enable it once
espressif updates the certificate.

Small changes to the example configuration passed
to the `eth_main` function, reflecting the
changes made on the w5100 driver.

Updated sdkconfig and the examples to the
espressif/esp-idf@1067b2870 revision of IDF.